requirements/local.txt
-r ./base.txt

# Debugging
ipdb<=0.13.9

# Tools
django-extensions<=3.2.0

# Testing
mypy<=0.971
pytest<=7.1.2
pytest-sugar<=0.9.5
pytest-django<=4.5.2
factory-boy<=3.2.1

# Code quality
flake8<=5.0.4
